Support parallel dynamic gufuncs
