+++ numba/np/ufunc/parallel.py
@@ -210,7 +210,7 @@ def build_ufunc_wrapper(library, ctx, fname, signature, cres):
 
 class ParallelGUFuncBuilder(ufuncbuilder.GUFuncBuilder):
     def __init__(self, py_func, signature, identity=None, cache=False,
-                 targetoptions={}, writable_args=()):
+                 is_dynamic=False, targetoptions={}, writable_args=()):
         # Force nopython mode
         targetoptions.update(dict(nopython=True))
         super(
@@ -220,6 +220,7 @@ class ParallelGUFuncBuilder(ufuncbuilder.GUFuncBuilder):
             signature=signature,
             identity=identity,
             cache=cache,
+            is_dynamic=is_dynamic,
             targetoptions=targetoptions,
             writable_args=writable_args)
